Move and rename update_fee_currencies
Since it does not have anything to do directly with the CeloBlockEnv, I
moved it to fee_currency_context.rs.
